requirements.txt
crewai
langchain_openai
duckduckgo-search
ipython
black
streamlit
openai